clang-tidy: Suppress `performance-enum-size` check warning
See: https://clang.llvm.org/extra/clang-tidy/checks/performance/enum-size.html

+++ test/mp/test/foo.h
@@ -21,7 +21,7 @@ struct FooStruct
     std::vector<bool> vbool;
 };
 
-enum class FooEnum : int { ONE = 1, TWO = 2, };
+enum class FooEnum : uint8_t { ONE = 1, TWO = 2, };
 
 struct FooCustom
 {
